+++ src/esrs/postgres/projector.rs
@@ -4,6 +4,21 @@ use uuid::Uuid;
 
 use crate::{AggregateManager, StoreEvent};
 
+/// This enum is used to instruct via [`Projector::consistency`] function which guarantees to have
+/// while projecting an event in the read side.
+/// - [`Consistency::Strong`] means that the projected data will be always available in the read
+///   side. In the actual default store implementation it implies that if a strong consistent
+///   projector fails the event will not be stored in the event store and the transaction rollbacks.
+/// - [`Consistency::Eventual`] means that there are no guarantees for the projected data to be
+///   persisted in the read side. In the actual default store implementation it implies that if an
+///   eventual consistent projector fails that event is stored anyway but nothing will be persisted
+///   in the read side. If no other projector fails the event will be stored in the event store with
+///   all the other projections and the transaction will be committed.
+pub enum Consistency {
+    Strong,
+    Eventual,
+}
+
 /// This trait is used to implement a `Projector`. A projector is intended to be an entity where to
 /// create, update and delete a read side. Every projector should be responsible to update a single
 /// read model.
@@ -12,6 +27,14 @@ pub trait Projector<Manager>: Sync
 where
     Manager: AggregateManager,
 {
+    /// This function could be used to instruct the [`Projector`] about its the [`Consistency`] level.
+    ///
+    /// It has a default implementation that returns [`Consistency::Strong`]. Override this function
+    /// to change its [`Consistency`] level.
+    fn consistency(&self) -> Consistency {
+        Consistency::Strong
+    }
+
     /// This function projects one event in each read model that implements this trait.
     /// The result is meant to catch generic errors.
     ///