+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   in the `AggregateManager` the `EventStore` associated type as `Box<dyn EventStore<Manager = >>` (with `Send` + `Sync` 
   bounds).
 - [[#115]]: Added `apply_events` to `Aggregate` with default implementation.
+- [[#122]]: 
+  - `Consistency` level enum.
+  - `consistency` function to `Projector` to instruct the `EventStore` on the persistence guarantees with default 
+    implementation returning `Consistency::Strong`.
 - [[#123]]: Added `postgres` documentation in docs.rs with `package.metadata.docs.rs` in `Cargo.toml`. Improved 
   modules documentation.
 
@@ -136,6 +140,7 @@ Refer to: [#107], [#108] and [#109]
 [0.6.2]: https://github.com/primait/event_sourcing.rs/compare/0.6.1...0.6.2
 
 [#123]: https://github.com/primait/event_sourcing.rs/pull/123
+[#122]: https://github.com/primait/event_sourcing.rs/pull/122
 [#117]: https://github.com/primait/event_sourcing.rs/pull/117
 [#115]: https://github.com/primait/event_sourcing.rs/pull/115
 [#114]: https://github.com/primait/event_sourcing.rs/pull/114